Initial folder and code restructure
