+++ src/Explorer/QueryCopilot/Shared/SamplePrompts/SamplePrompts.test.tsx
