+++ src/Explorer/QueryCopilot/V2/Bubbles/Sample/SampleBubble.test.tsx
@@ -1,0 +1,10 @@
+import { shallow } from "enzyme";
+import React from "react";
+import { SampleBubble } from "./SampleBubble";
+
+describe("Sample Bubble snapshot test", () => {
+  it("should render ", () => {
+    const wrapper = shallow(<SampleBubble />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
