+++ src/Explorer/QueryCopilot/V2/Bubbles/Welcome/WelcomeBubble.test.tsx
@@ -1,0 +1,10 @@
+import { shallow } from "enzyme";
+import React from "react";
+import { WelcomeBubble } from "./WelcomeBubble";
+
+describe("Welcome Bubble snapshot test", () => {
+  it("should render ", () => {
+    const wrapper = shallow(<WelcomeBubble />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
