+++ src/Explorer/QueryCopilot/V2/Footer/Footer.test.tsx
@@ -1,0 +1,10 @@
+import { shallow } from "enzyme";
+import React from "react";
+import { Footer } from "./Footer";
+
+describe("Footer snapshot test", () => {
+  it("should render ", () => {
+    const wrapper = shallow(<Footer />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
