+++ src/Explorer/QueryCopilot/V2/Header/Header.test.tsx
@@ -1,0 +1,10 @@
+import { shallow } from "enzyme";
+import React from "react";
+import { Header } from "./Header";
+
+describe("Header snapshot test", () => {
+  it("should render ", () => {
+    const wrapper = shallow(<Header />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
