+++ src/Explorer/QueryCopilot/V2/Modal/WelcomeSidebarModal.test.tsx
@@ -1,0 +1,10 @@
+import { shallow } from "enzyme";
+import React from "react";
+import { WelcomeSidebarModal } from "./WelcomeSidebarModal";
+
+describe("Footer snapshot test", () => {
+  it("should render ", () => {
+    const wrapper = shallow(<WelcomeSidebarModal />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
