+++ src/Explorer/QueryCopilot/V2/Sidebar/QueryCopilotSidebar.test.tsx
@@ -1,0 +1,10 @@
+import { shallow } from "enzyme";
+import React from "react";
+import { QueryCopilotSidebar } from "./QueryCopilotSidebar";
+
+describe("Footer snapshot test", () => {
+  it("should render ", () => {
+    const wrapper = shallow(<QueryCopilotSidebar />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
